Remove unneccessary complexity and duplicate code
_FroupSync is the parent class used for building all the consent sync
classes. There are currently two syncs which are using it to build a
sync class, consent_gsuite and consent. Both these syncs use their own
child class to implement the method pe2accs, but the implementations
are identical, meaning that we should move that implementation to the
parent class. When this has been done, ADGSuiteConsentsync and
ADo365consentsync are empty and should be deleted.

Issue: CRB-3064