Remove use of AuthPGP for uio guest accounts
Previously, the uio guest accounts got a new password when the account
was released into the pool of free guest accounts. This password was GPG
encrypted and stored in account_authentication using
Cerebrum.modules.AuthPGP. This password was fetched and decrypted when
reserving the guest account for use.

In an attempt to get rid of AuthPGP, these guest accounts now get a new
password when they are reserved for use, and the password is simply
remoevd when releasing the account back into the pool of free guest
accounts.

Issue: CRB-1039