Fix entity_contactinfo_* case-sensitivity for contact_type and contact_value
Fixed the contact_value to not get lowercased, as mentioned in the issue we want to store the verbatim contact_value. In addition I've added the .lower() to the contact_value when checking for previously added values that are identical to the one we attempt to add.

For contact_type the issue mentions EMAIL has to be typed in to bofh with captial letters for the command to work. This will be the same for MOBILE and the other constants. To fix this I've added a .upper() to the _get_constant() function, making sure the type_str is uppercase when getting it.

Issue: CRB-2024