Change params to avoid future bugs and fix command order (optionals last)
Changed the params to DistributionGroup.new() to *args and **kwargs so it is more resistant
to changes in the future, to avoid similar bugs.
Removed check for manual group when setting display name so we are able to change display name on automatic groups as well.
Changed the command options to work properly, with optionals last and default values.

Issue: CRB-3024