Rewrite test_utils/test_gpg.py with pytest
This commit re-implements the 'test_core/test_utils/test_gpg.py'
unittest module as a pytest module 'test_gpg/test_utils_gpg.py'.

Both the rewrite and the move is done in order to make use of the gpg
fixtures that are already in place.  This removes the pre-requisite of
having a working `cereconf.GNUPGHOME` set up with a given key before
running the tests.