Rename argument of a method
The change_type argument of the log_change method of event_publisher does
not match the same argument it calls super for. This means that if you use
call the method using keyword arguments it misses the argument, and you
get an error. This commit renames it to match.

Issue: CRB-3019