Rewrite ad2 unit tests as pytests
This commit rewrites the unittest module for ad2 as a pytest test
module. The gpg test is skipped as the functionality is no longer
implemented as the test expects and will fail -- also I can't be
bothered to figure out a good way to use the test_gpg fixtures from this
test subdir, or find a way to organize these tests to let them share
fixtures.