Rework code for distribution lists in CIM
The original base class did not create distribution lists. This commit
adds a call to the create_dist_lists method in get_person_data, and an
empty create_dists_lists method that is meant to be overridden by
instance specific methods if one wants to add a distribution list.

The UiT method now stops clearing and repopulating the same person object.
In addition we add the create_dists_lists method for the UiA datasource.

Issue: CRB-3028