Add a new hidden command used to display more group in brukerinfo.
This command makes some assumptions on how "person" groups are exported..
Most automatic groups are "person" groups. When these are exported we
export the primary user, and we would like to display these in
brukerinfo.

To enable this, we return the person memberships as memberships on the
primary account. For non-primary accounts, only the accounts groups are
returned.

Issue: WOFH-27