Fix a "sleeping" bug
This part of the code does work at the moment, but it is not very
robust. It hinges on the fact that 'assignments' and 'roles' are
the only keys in the data from SAP that contain the value
{"__deferred": {"uri": "..."}}. If any key other contained that
value, this would break because filter_param is not defined.

Issue: CRB-3048