Validate usernames in the relevant classes
Only PosixUser requires that the name of a new account does not collide
with the name of an existing group. That requirement exists so that a
posix user can get a personal group with the same name. It is not
necessary to check whether the username is taken by group when
populating a regular account.