Fix issue with employees on temporary leave
If an employee is on temporary leave, their accounts and permissions in
Cerebrum should be suspended until they return. Since we were previously
looking at an assignments' 'originalHireDate' instead of
'effectiveStartDate', we would still include assignments with
'effectiveStartDate' set in the future. This should fix the issue where
we wrongly include some persons on temporary leave.