virthome: refactor and format "group invite_moderator"
This improves the readability of group_invite_moderator() by making it
explicit which the included fields are in the response body.  Before,
a single dictionary was created and updated in different if-conditions,
making it unnecessarily hard to reason about what the final product would
be, especially because try-except is being used to control code flow.

Fixes: CRB-250