Add LdapConfig class as alternative to ldapconfig
This commit adds an LdapConfig class, which can be used to replace
`ldapconf()`.  The idea is to force `cereconf.LDAP_*` to be fetched on
script init and passed into the relevant LDAP export class - which then
can be used to provide alternative config from other locations than
`cereconf`, or to avoid hard-coding names like `LDAP_ORG`.