virthome: document and format "user recover_uname"
The "user recover_uname" bofh command is unique in that it returns
a plain JSON Array response body.  The format suggestion provided
by this patch works in pybofh but fails in jbofh because it
unconditionally coerces the body to a java.util.HashMap.

There is little we can do about this without either altering the
payload data structure or patching jbofh.  Changing the data format
is a hard limit due to the way the WebID web frontend is deployed,
so the best likely course is to make the format suggestion work in
pybofh and address the protocol parsing in jbofh later.

Fixes: CRB-250