virthome: document and format "spread list"
Provides a plain one column text list of spreads that applies to a
specific entity, e.g. an account or group.

This formatting shares the same problem as "user recover_uname" in
that it only works with pybofh due to a likely bug with the way
jbofh handles returned arrays.

Fixes: CRB-250