pwcheck: make PasswordNotGoodEnough Unicode compatible
PasswordNotGoodEnough and exceptions derived from it are commonly
constructed with localised messages from gettext(3).  These messages
often contain wide code points and are always unicode typed.  This
patch makes the exception classes provided by Cerebrum.modules.pwcheck
Unicode compatible.

BaseException, which all classes that inherit from Exception
ultimately inherits from, provides three methods for accessing the
message they are constructed with:

  (1) BaseException(T).message -> T
  (2) BaseException(T).__str__() -> str
  (3) BaseException(T).__unicode__() -> unicode

Unfortunately, BaseException makes the assumption that T is always
a byte string.  This is exemplified by its __unicode__() implementation
which coerces the result of __str__() without consideration for
encoding.  Put in other words, the expectation is that T must be a
byte string and that whatever data it holds must be encoded prior
to initialisation.

We are able to demonstrate this by constructing an exception with
a unicode message:

	>>> str(BaseException(u"æøå"))
	Traceback (most recent call last):
	  File "<stdin>", line 1, in <module>
	UnicodeEncodeError: 'ascii' codec can't encode characters in position 0-2: ordinal not in range(128)

Normally one might consider using the @python_2_unicode_compatible
decorator from future.utils which is designed to solve precisely
this problem.  Unfortunately, it overrides __str__() and uniquely
for exception types, __unicode__() operates on the underlying message
argument passed in during construction, and not as you might except,
the result of a user-implemented __str__():

	https://github.com/python/cpython/blob/2.7/Objects/exceptions.c#L123-L125

This means that in order to deal with the problem at hand, we cannot
override PasswordNotGoodEnough.__str__() to return an encoded byte
string as we might do under normal circumstances.  Additionally,
overriding __str__() has the downside that it would not be
forward-compatible with Python 3, where str is Unicode compatible.

Instead of burdening caller with the responsibility to encode the
input, we can however change PasswordNotGoodEnough's constructor
to gain a little extra safety.  This means we instead have decode
T on unicode(PasswordNotGoodEnough(T)).  This should be fine, since
__unicode__() is not observed when run under Python 3.

A final word of warning is that this patch is a workaround for the
Unicode problems of Python 2 and that all this code will be superfluous
once Cerebrum runs under Python 3.  However, in an advance to get
to that point we need the codebase to run under both interpreter
versions.

Fixes: CRB-3223