pwcheck: install gettext helper with unicode=True
Whilst the Python documentation advises to use unicode=1, its
internals expect the argument to be boolean:

	https://github.com/python/cpython/blob/2.7/Lib/gettext.py#L317

This makes no practical difference, however, since 1 is a truthy
value.  It does however lessen the cognitive load when reading the
code as the reader does not have to worry whether the number 1 holds
any specific meaning.