utils: introduce phone interface
Collect phone number related functionality such as parsing, validation,
and formatting in an API internal to Cerebrum.  This enables us to
more easily exchange the third-party dependency libphonenumbers in
the future.

The Cerebrum.utils.phone API is similar to that of the Python port
of libphonenumbers, except it assumes a few defaults that are
characteristic of Cerebrum:

	from Cerebrum.utils import phone

	numobj = phone.parse("22 85 50 50")
	numobj.country_code)  # 47
	numobj.national_number  # 22855050L

	phone.is_valid(numobj)  # True
	phone.format(numobj)  # u"+4722855050"

As we can see from the above example, it assumes a default region
from cereconf and E164 as the default formatting.

The tests for this feature are not intended to exhaustively test
phone number parsing, which is decidedly out of scope for this work.

Fixes: CRB-2805