contrib.no.uit: use Cerebrum.utils.phone
Cerebrum.utils.phone is a wrapper for phonenumbers with some Cerebrum
specific conveniences added, such as defaulting to the E.164
formatting.

Notably, this patch changes fetch_mobile.py from looking at whether
the phone number is possible to whether it is valid.  A validity
check of the putative phone number is more exact (and more expensive)
than judging if it is possibly a phone number, because it uses a
full validation of the number including region using length and
prefix information.

Fixes: CRB-2805