bofhd: refactor BofhdRequestHandler._send_response()
_send_response() relied on truthiness checks of the response argument,
which could variably be None or a tuple, potentially triggering
different behaviour depending on the value of the tuple.  This makes
_send_response()'s behaviour more deterministic, as we should only
have to encode the XML-RPC response if indeed have something to
respond with.

For accountancy it also helps to introduce different body, payload,
and response variables so that response is not continuously mutated.