Cerebrum.utils: introduce unicodestring module
The Cerebrum.utils.unicodestring module introduces various routines
for Unicode string manipulation.

At present it exposes functionality for stripping and replacing
entire subsections of code points associated with a particular
Unicode code point label.  These are known  as "categories" in
Python's standard library unicodedata, and as such the terminology
from there is carried on.

strip_control_characters() is a shorthand for removing control
sequences from strings, which we see become useful for bofhd.
Control sequence characters, generally speaking, are unprintable
and affect cursor movement in visual (graphcial) terminals.  They
include the likes of sounding an alarm, printing a tabulator, and
aligning layout using line feeds, carriage returns, and vertical
spacers.

bofhd serves mainly terminal-based clients (jbofh, pybofh) that are
subject to all the laws and physics of 1970s terminal emulators.
As such, a control character for backspace (BS) that erases the
previous character will, for the user, be experienced as data loss:

	abc\x08

Will be printed to the terminal as:

	ab

Further, xmlrpclib is apparently unable to encode certain control
sequence characters because it rather worryingly operates on the
byte-character level instead of at the code point level.  This
occasionally causes the produced XML to be invalid.

Ideally the input to the Cerebrum would undergo stricter sanitisation
before it gets committed to the database. But we are left to deal
with reality, and that is that we at least need to wash the data
before printing to the primary administration interaction point of
Cerebrum, that is bofh.

Fixes: CRB-1347