Fix KeyError issue in uit group ldif export
This commit fixes an issue in the UiT-copy of generate_groups_ldif.
This issue occurs when an *expired group* exists with LDAP spread and
members, as group.search() by default excludes expired groups, but
group.search_members() includes members of expired groups.

The short term fix is to simply ignore memberships of expired groups
when building the pickle cache, as those groups aren't exported anyway.
The next step would be to re-write the three generate_groups_ldif.py and
optimize the group and member search queries to exclude any unwanted
data.