feide: fix service_list formatting
The jbofh client is not too fond of the %d formatting code and it
is safer to use %i which is native to Java.

The feide_id may be either a number or a (UUID) string.  The patch
adapts the formatting layout to take this into account.  UUID(4)s
are always 36 characters in length.

Fixes: CRB-3234