pip: pin werkzeug to v0.16.1
werkzeug is a dependency of flask-restplus, but flask-restplus does
not exactly define which version it requires.  This means the version
werkzeug version pip installs is incompatible with flask-restplus.

Before we can begin the task of upgrading flask-restplus we need a
working environment.  This specifies the exact werkzeug version
flask-restplus needs so that a fresh installation will work.

Fixes: CRB-3254