Add hr_import logic
This commit adds a hr_import consumer script, along with modules to
configure and run the import itself.

TODO: The hr_import design is currently a bit *over-abstracted* - we may
want to simplify some parts, but this will be easier to do once we get a
second hr import implementation to work with.

Issue: CRB-3202