Fix an issue with `can_alter_group` at UiA
At UiA `can_alter_group` is always true if a given operator is an admin
for a given group.  This also means that anyone who is an admin for any
group should be shown the command.  This commit fixes a bug where we may
end up calling a permission check for a specific group when running with
`query_run_any=True`.  This is never the case.