Add custom hr import model representation.
This commit implements a custom `repr()` result for models in the
`hr_import` module, which should make them more helpful when included in
log messages and exceptions.

Fixes an issue with `HRAffiliation.is_active()` where we might get an
`OverflowError` when calculating the start and end cutoff dates.

Issue: CRB-3325