Change hr-import secrets config
This commit adds use of `Cerebrum.config.secrets` to
`Cerebrum.modules.amqp.config`, and changes the `Secret` setting to be
an actual setting, and not a `Namespace`.

This change simplifies the config, and makes it easier to provide
default values.  E.g. providing username="guest", password="guest" as
default credentials for `Cerebrum.modules.amqp` makes it easier to get
a fully local test environment with a RabbitMQ docker container up and
running for development purposes.

Issue: CRB-3369