Change orgera_assignments primary key
This commit changes the orgera primary key to include source-system, and
removes the (person, ou, sko) contstraint.

It's not immediately clear what the primary key *should be*.  We've
assumed that the assignment_id is unique in a given source system, but
that is not given. Maybe two persons could share an assignment? If so,
the key should be (source, assign, person).

Issue: CRB-3322