Change person_otp_secret.created_at to updated_at
This change renames a column in person_otp_secret from created_at to
updated_at, and adds a trigger for automatic update of the column value.

Additionally, the column is no longer possible to explicitly set using
the db update functions in `Cerebrum.modules.otp`.  It is now set only
through its default value (`now()`) on INSERT, and through its trigger
on UPDATE.