Fix issue with finding forward_to addr for sysadm
This commit fixes an issue with finding the appropriate forward_to email
address for a given sysadm target account.

Also adds util function for fetching primary email addresses for a given
target type.  This will be useful when figuring out which email address
to use for sysadm accounts.

Issue: CRB-3274