Fix minor issue in bofhd auth
This commit fixes issues with the common and tsd implementations of
`can_alter_group`:

in common auth: No need to check if superuser when `query_run_any=True`
permits everyone to see the command.

in tsd: Fixes an issue where we check for *specific* group auth
permissions when running with `query_run_any=True`.  The effective
result is the same (command hidden if not superuser or group admin for
any groups), but we don't get a traceback in the logs.