Add option to disable feide sms mfa labels
This commit adds a `LDAP_PERSON[norEduPersonAuthnMethod_labels]` option.
The option is `True` by default, but can be set to `False` to disable
the `label` part of sms auth entries in `norEduPersonAuthnMethod`.

If no label is given in a sms entry, Feide defaults to showing the last
two digits of the phone number.  Duplicate phone numbers are omitted.

Issue: CRB-3941