Rewrite hr-import mapper and models
This commit rewrites the HR-import to no longer use the objects in
Cerebrum.modules.hr_import.models, and removes these objects as well.

The overall design is simplified a bit, and it's clearer from just
reading the the mapper code alone to see what data is imported.

Other changes/fixes:

- Having *two* affiliations of the same type, at the same OU, but at
  different periods are now supported

- Persons marked as deceased are *not* updated from dfø.  This prevents
  any issues where a deceased person is suddenly re-activated because
  *one* source system isn't updated with this information.

Issue: CRB-3968