Fix issue with `ou info` (address country)
If an OU address contains a country field, the `ou info` bofhd command
will fail.  This is because we try to use the country code constant
number in the output, when we *should* look up the country code string.

This doesn't really happen much, as pretty much only old, invalid org
units contains country codes...