Fix error and error message in matcher import util
This commit fixes an issue where we get a `NotFoundError` (indicating
that the given object doesn't exist) when one of the external-id search
types given to various `Cerebrum.modules.import_utils.matcher` search
utils doesn't exist in the database.

This commit ensures that all search types given actually exists in the
database before we search.  If this isn't the case,
`ConstantsBase.get_constant` will raise a `LookupError` with a more
appropriate error message.

Issue: CRB-3975