Simplify GregImport.CONSENT_GROUPS behaviour
This commit changes how GregImport.CONSENT_GROUPS works.  Rather than
mapping consents to callbacks for setting/removing consents, it now maps
consents to group templates.  This should reduce code-duplication when
adding consent groups in subclasses.

We may want a generic `sync_consents` method in the future with the old
callback functinality (especially if we decide to store consents using
the consent module), but this solution is simpler for now...

Issue: CRB-3975