Fix odd issue with spread-expire in `greg_users`
`add_spread` always sets the spread expire date to *today*.  This is
odd, but probably works because UiT only calls `add_spread()` if the
spread is actually missing.

Since most imports are fullsyncs, this would work for UiT because the
*next* import would fix the expire-date.  However, for event-driven
imports, we'll need to set the actual spread-expire dates *after*
spreads have been added.

Issue: CRB-3976