Fix a regression in `group personal`
This commit fixes an issue in the `group personal` bofhd command.  This
command calls `_group_add` -> `_group_add_entity` to ensure that the
user is a member of its own personal file group.

When doing so, we call `_group_add` *without* an operator to ensure that
some permission checks *don't* run, but a change in `_group_add_entity`
requires an *operator* to be given.