Fix log messages in send_welcome_sms
This commit changes log messages related to skipping/giving up on
candidates for welcome sms.

Some of the previous log messages were not very helpful when debugging:
E.g. "Tagged new user %r with quarantine, skipping" sounds like we're
adding a quarantine to the account, when we're actually skipping the
account because of a quarantine.

The new messages all follow the formats:

- If we'll retry: "Skipping account %s (%d): <reason>"
- If we won't retry: "Giving up on trait %s for account %s (%d): <reason>"

Also deals with an issue where the script might crash if an invalid
selection trait is set (e.g. 'new_account' with a missing date).  This
would only happen if someone adds the trait using e.g. `trait set` in
bofh, but we should try not to crash if that happens.