Change bofhd session timestamps to include tz
This commit changes bofhd session timestamps to no longer use
`mx.DateTime`.  This is done in two distinct ways:

1. By changing code to (in theory) handle both `mx.DateTime` and
   `datetime.datetime` values

2. By changing the timestamp columns to be represented with tz, which
   should cause our database driver to return tz-aware datetime objects,
   rather than `mx.DateTime` objects.

In the long run all timestamps should probably be stored tz-aware, but
that is a problem for another day.  The bofhd module is simple to
update/migrate, as timestamps aren't really used or looked at outside
the bofhd session module.

Issue: CRB-3445