Change API datetime resolution
This change removes sub-second resolution in responses that use the
naive, localized iso datetime field.

This is done to prevent any changes to the "format" of this field.  Our
recent changes from `mx.DateTime` to `datetime.datetime` means that the
value passed to this serializer may include sub-second data.

This isn't really a format change, as these fields have always included
sub-second data (in ISO-8601 format) *if provided*, it's just that we
never provided the data...

However, we can't be sure that all our consumers *actually* parse these
dates correctly, so it's best to prevent any changes for now.

Issue: CRB-3433