Speed up the UiA SAP import
This commit re-implements the function that clears out-of-date data from
persons in the UiA SAP-import.  This is done by selecting *only* the
persons that actually *have* information to remove, rather than
procesing *all persons in Cerebrum*.

The speed-up is significant in test, reducing the total import time
from ~10m to ~1m 30s.

Issue: CRB-4070