Remove defunct, generic batch import
This commit removes `Cerebrum.modules.process_entity`, as well as the
script to trigger this functionality.  This was a somewhat generic
framework for batch import and user creation, but was only ever in use
at ØFK (and maybe Giske?).  Any future attempts at a generic framework
for similar behaviour should not be focused around *batch processing*.

Issue: CRB-4072