Remove `Cerebrum.modules.abcenterprise` and friends
This commit removes the `Cerebrum.modules.abcenterprise`, along with the
script that would use this framework.  The ABC Enterprise data format is
obsolete, and we no longer have batch imports in use that that needs to
process this format.  Note however, that the standalone script to
*export* data on this format is still in use
(`contrib/no/export_abc.py`).

This commit also removes the `Cerebrum.extlib.doc_exception`.  This was
only in use by `Cerebrum.modules.abcenterprise`.  Similar behaviour is
also implemented in `Cerebrum.Errors`.

Issue: CRB-4073