Add a person search endpoint to the rest api
Introduces the search/persons/affiliations endpoint to the rest api,
which can be used to get the person ids of anyone with one of the
affiliations supplied to the affiliations query arg. The endpoint is not
listed in the api docs since we don't want this to be the permanent
solution for listing persons.