Add lookup to catch upcoming dfo assignments
This commit adds an extra employee lookup, in order to catch new
assignments that will start in the near future.

The DFØ API doesn't show previous/upcoming _main assignments_, so
whenever this changes for a person, there is an overlap where we
don't consider the person as an active employee.

This commit adds an extra employee lookup to check if there are
additional assingments three days from now.  The number is kind of
arbitrary, but should catch most of the issues we encounter.

Issue: CRB-4064