Fix `person affiliation_remove` aff parameters
This commit fixes an issue in the bofhd command 'person
affiliation_remove' where operators are unable to delete a manual aff
because another, inaccessible aff of the same type on the same org unit
exists.  The root of the issue is that our selection criteria
(person-id, ou-id, affiliation) can match more than one affiliation.

The fix is to allow more fine grained selection criteria. If given the
same arguments as before (person, ou, aff), the same affiliations are
selected, *but* the affiliation argument can optionally include
affiliation status, and an additional, optional source system parameter
can be added.  Examples:

- `person affiliation_remove rektor 900000 ANSATT` (as before)
- `person affiliation_remove rektor 900000 ANSATT/tekadm`
- `person affiliation_remove rektor 900000 ANSATT MANUAL`
- `person affiliation_remove rektor 900000 ANSATT/tekadm MANUAL`

Previously, if more than one affiliation was matched, *and* the operator
has access to delete all of them, only one would be deleted (the first
selected by `list_affiliations()`) .  This is also changed - all matched
affiliations are removed as long as the operator has access to this.

Issue: CRB-4067