Remove duplicate affiliation definitions
This commit tries to reduce the number of duplicate affiliation
definitions, by re-using parent definitions.

Changes:

- Common affs and aff statuses has been moved up to superclasses, for
  re-use.

- ConstantsUniversityColleges is now a subclass of ConstantsHigherEdu.
  This was enforced through config for everyone using
  ConstantsUniversityColleges anyway.

- Some constants are now available through *multuple* attribute names.
  This is to ensure backwards compatibility.

This change won't add or remove any affs or aff statuses in any current
environments, *but* some aff/status descriptions will be updated.

This change *shouldn't* require us to update `CLASS_CONSTANTS`, but it
does.  This is because the current configs adds constant classes in the
wrong order (i.e. parent _before_ child) - and this needs to be fixed
because we now enforce order through inheritance.

Issue: CRB-4104