Remove bytes support in `legacy_read_password`
`legacy_read_passwords` tries to read bytes directly from file by
default in both PY2 and PY3.  This causes issues in PY3, as we generally
try to process the file content as *text*.

There really isn't any use case for this function to read and return
bytes.  If this needs to be supported, then:

1. We'd need to ensure most (if not all) calls to `legacy_read_password`
   includes an encoding so that the result is unicode/text

2. We'd need to rewrite the function to properly support bytes.

Here we take the easy way out, and default to decoding the file content
as utf-8.