Remove two unmaintained import/migrate scripts
This commit removes `contrib/migrate/import_netgroups.py` and
`contrib/migrate/import_unames.py`.  These scripts aren't relevant, and
it's a waste of time to maintain them for future use.  We'd probably
want something with better validation and normalization if this is ever
needed again in the future.

Issue: CRB-4096