Rewrite generate_accounts.py
This commit rewrites the `generate_accounts.py` script.  This change is
done in order to:

1. Make the script a bit easier to follow
2. Fix some issues with potential conflicts in input arguments
3. Fix a bug where affiliation status won't work as an inclusive
   selection criteria (e.g. create accounts for ANSATT/tekadm wouldn't
   work)
4. Fix an issue where the script could fail with CEREBRUM_NOMX=1

As a bonus, the script is now a lot faster, as we don't need to do
lookups for persons that already has accounts.

Also, the script logs a bit less noise.  CLI/account settings are logged
at startup, rather than once for each account update, and we no longer
log persons that aren't processed.

Issue: CRB-3429