Remove `mod_entity_expire`
This commit removes the `mod_entity_expire` and
`Cerebrum.modules.entity_expire` modules.

It seems this module was only ever in use at UiT, even if it can be
traced back a bit to other environments as well.  A lot of the test
environment setups have included this module for some reason.  None of
the actual databases have the changelog constants, which *should* mean
there are no references to the module from changelog or other places.

Issue: CRB-4152