Fix an issue with parsing date and time values
This commit fixes an issue where `Cerebrum.utils.date` parsers throws
`NotImplementedError` when a valid, but unsupported ISO8601 value is
given.  This typically happens in bofhd commands, where the
NotImplementedError results in an "unknown error" being given to the
user.

The fix changes the error type to a `ValueError`, which is what we'd
expect a parser to throw when given an unparseable value.