Remove duplicate constant type, and fix script
This commit removes the duplicate `AuthConstants` constant type, which
is just a duplicate of the `BofhdAuthOp` code.

Also cleans up a script for listing empty groups.  This script *used* to
refer to the group-owner bofhd op code, so had a superfluous
`AuthConstants` reference.  The script took too long to test, so it's
been optimized a bit - now runs about 50x faster.

Issue: CRB-4202