Fix issue that prevents access to `user history`
This commit fixes an issue that prevents access to user history.  This
regression was caused by changing string literals to unicode objects in
the bofhd auth module, combined with a bug in `db_row.row.get()`.

`db_row.row.get()` works differently from `db_row.row.__getitem__` when
dealing with string lookups:  When the actual key is a bytestring, and
the lookup key is a unicode object (and probably vice versa),
`db_row.row.get` results in a miss, and returns `None`.

The `get` and `_getitem_ implementation in `dict` and our new `records`
module are both consisten here, and both works as expected.  I.e.  `get`
and `__getitem__` both hits works when the actual key and lookup key are
mixed string types.

Issue: CRB-4213