Remove `Cerebrum.utils.context`
This commit removes the entity context module.  It's not widely used, a
bit too complicated for what it does, and worst of all - uses a global
db connection!

The module was a nice idea, but would need some re-work to be really
useful, and we'd need to make it Python 3 compatible.