Fix an issue with `EntityGPGData._gpg_encrypter`
This commit fixes an issue where `Cerebrum.utils.funcwrap.memoize`
cannot be used to cache the gpg config (encrypter object), because the
method object (`self`) is unhashable in Python 3.

This (and similar issues) can easily be fixed by using a
`Cerebrum.utils.descriptors.lazy_property` as a drop-in replacement for
`property(memoize(func))`.

Issue: CRB-4225