Add better naming of logger threads
This commit adds better support for naming multiprocessing logger
threads, as well as default names for the logger threads.

If the optional `setproctitle` module is installed, we'll also set the
thread names using appropriate syscalls, so that they are available in
e.g. `top`, `ps`, and similar utils.  This is very useful when
debugging.