Remove broken script for populating target systems
`<cerebrum-hacks>/felles/clear-target-system.py` can be used to remove
these mappings when needed.

In order to re-populate these systems from scratch, we'll have to use a
db-client to run the queries and statements manually.  It'll probably
never happen, and if it does, it's still less work than fixing this
script.

Issue: CRB-3547