Re-write password generators
This commit re-writes the
`Cerebrum.modules.password_generator.generator` module.
PassphraseGenerator isn't really in use currently, but it *should* be
for e.g. NIH.

This is an attempt to clean up the code, and make it a bit more flexible
if we ever want to generate passphrases.

Future improvements:

- Move this module out of `Cerebrum.modules` (as it is required by
  `Cerebrum.Account`), and make it actually configurable if we want to
  use passwords or passphrases.

- Replace existing `make_passwd` overrides with config

- Deprecate `Account.make_passwd` and replace it with the configurable
  password generator

Issue: CRB-4241