Fix isse when moving hard-coded path back to disk
The previous commits fixed an issue where `user move hard_nofile` *kept*
the disk value when attempting to move a on-disk homedir to a non-disk
path.

This commit fixes a related issue, where moving a non-disk path value
back to an on-disk homedir with `user move nofile` doesn't reset the
homedir path.

Issue: CRB-659