Remove unused, broken db lock mechanic
This removes the `Lock`, `OraPgLock`, and the non-standard cursor/db
`acquire_lock` methods.

This doesn't seem to be used anywhere.  The `Lock`-init has a typo bug
that prevents it from being created, so it wouldn't work in its current
state.  The bug was fixed shortly after implementing the locks in 2015,
but re-introduced in 2018.

Issue: CRB-4273