Remove `Cerebrum.modules.synctools`
This module has never been used.  It was intended as a fullsync backup
for a potential event based Active Directory sync.  If we are to
implement a new sync, it would not, however, work like this.

This implementation has a few issues:

1. It doesn't actually work properly, and is not Python 3 compatible.

2. It relies on legacy `Cerebrum.modules.ad` settings.

3. It emits events to the Cerebrum exchange, which would send
   notifications to *all* subscribers.

4. It doesn't re-use the message formatting utils in
   `Cerberum.modules.event_publisher`, so messages could be incorrectly
   formatted.

It's probably easier to remove this module, so that we don't have to
maintain it.

Issue: CRB-4284