Remove obsolete hr-import mapper config
All business logic related to the mappers have been moved to custom
mapper classes and custom import classes.  This simplifies the config a
bit, and (maybe) makes the setup steps a bit easier to follow.  There's
a lot more opportunity to simplify these parts of the hr-import, though.

Issue: CRB-4065