Refactor populate-azure-groups
This commit adds a helper module for maintaining automatic groups,
`Cerebrum.group.group_sync`.  The `populate-azure-groups.py` script is
refactored using this module, for a substantial speed-up.

In test, this script now runs in ~1 minute (rather than ~25 minutes)
when there are few changes.  The overhead for changes (adding/removing
members) should be the same.